used polldatamodel and moved write to page from component
