+++ packages/nextjs/components/poll/PollComponent.tsx
@@ -5,38 +5,10 @@ import HoverBorderCard from "../card/HoverBorderCard";
 import CountrySelector from "../country_picker/CountryPicker";
 import { COUNTRIES } from "../country_picker/countries";
 import { SelectMenuOption } from "../country_picker/types";
-import { PubKey } from "@se-2/hardhat/domainobjs";
+import { PollData } from "./PollDataModel";
 import { LuCross } from "react-icons/lu";
 import { MdEdit } from "react-icons/md";
 import { RxCross2 } from "react-icons/rx";
-import { useChainId } from "wagmi";
-import { useScaffoldContractWrite } from "~~/hooks/scaffold-eth";
-import { contracts } from "~~/utils/scaffold-eth/contract";
-
-/**
- *
- * KP 1
- * [✓] Public key: macipk.31c3b9d36aed8ab9490d7468cb581669ff2774791c7fcefc4cdd07bd009a019d
- * [✓] Private key: macisk.79a8acc8572683cd2f6a212b7ac096f66b39dbad6ceb11634a1b0d7584df4856
- *
- * KP 2
- * [✓] Public key: macipk.f48bc9877d99ba9482ababaa53f31d492d84bf105e0c8b3fc60edbf52ed7f0a3
- * [✓] Private key: macisk.578e007d0eb7625624c00e898379bc2b3e96af59dffa7212eb2a5ae100f24c73
- *
- */
-
-const coordinatorPubKey = PubKey.deserialize("31c3b9d36aed8ab9490d7468cb581669ff2774791c7fcefc4cdd07bd009a019d");
-
-function keyToParam(key: PubKey): { x: bigint; y: bigint } {
-  const p = key.asContractParam();
-  return { x: BigInt(p.x), y: BigInt(p.y) };
-}
-
-export interface PollData {
-  title: string;
-  options: string[];
-  country: SelectMenuOption;
-}
 
 const PollComponent: React.FC<{ onPollDataChange: (data: PollData) => void }> = ({ onPollDataChange }) => {
   const [pollData, setPollData] = useState<PollData>({ title: "Dummy Title", options: [""], country: COUNTRIES[0] });
@@ -48,7 +20,6 @@ const PollComponent: React.FC<{ onPollDataChange: (data: PollData) => void }> = 
     setCountry(selectedCountry);
     setPollData({ ...pollData, country: COUNTRIES.find(c => c.value === country) || COUNTRIES[0] });
   };
-  const chainId = useChainId();
 
   const handleAddOption = () => {
     setPollData({ ...pollData, options: [...pollData.options, ""] });
@@ -77,33 +48,6 @@ const PollComponent: React.FC<{ onPollDataChange: (data: PollData) => void }> = 
     newOptions.splice(index, 1);
     setPollData({ ...pollData, options: newOptions });
   }
-
-const { write, data } = useScaffoldContractWrite({
-    contractName: "MACI",
-    functionName: "deployPoll",
-    args: [
-      90000n,
-      {
-        intStateTreeDepth: 1,
-        messageTreeSubDepth: 1,
-        messageTreeDepth: 2,
-        voteOptionTreeDepth: 2,
-      },
-      keyToParam(coordinatorPubKey),
-      (contracts && contracts[chainId] && contracts[chainId]["Verifier"]?.address) || undefined,
-      (contracts && contracts[chainId] && contracts[chainId]["VkRegistry"]?.address) || undefined,
-      false,
-    ],
-  });
-
-  console.log(chainId);
-  console.log(data);
-
-  const createPoll = () => {
-    console.log(pollData);
-
-    write();
-  };
 
   return (
     <div className="max-w-2xl mx-auto p-6 shadow-2xl rounded-md  bg-gradient-to-r from-[#19244F]  to-[#181436]">
